pyproject.toml
[project]
name = "authenticationsystem"
version = "0.1.0"
description = ""
authors = [
    {name = "lorna88",email = "lorna_88@mail.ru"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=6.3.0,<7.0.0)",
    "ipython (>=9.12.0,<10.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-stubs (>=3.16.9,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
